#59: Don't print schedule task 'reasons' in mod log messages
For now the internals of the scheduler are still up in the air, and I don't know if we want "reason" strings attached to every task in the long run, _but_ what's definitely true today is that they're not designed to be user-facing output, and as long as they're being created at all there's no reason to exclude them from the dev logs.

The primary goal here is to make Izzy's remove role mod message less redundant, since the "reason" strings for those tasks today are just redundant. I also tried to make add role consistent.