.help output tweaks for alternate names
Fixes https://github.com/Manechat/izzy-moonbot/issues/58

I noticed the version of Izzy on prod today contains one of the three changes I was after in that issue, but none of the others, which is ironically more inconsistent than before, and I found none of them were on `mane`. Not sure what's going on but since I got that far I figured might as well PR this myself.