Make the pressure embed correct by distinguishing between "old" pressure before and after decay is applied
I believe this corrects the mistake you were explaining to me on Discord after my previous PR.

This change to IncreasePressure might change the timestamp it saves to the user, but this looks to me like a drive-by bugfix. The existing code has IncreasePressure fetches the current time before calling GetAndDecayPressure, so when IncreasePressure sets the timestamp it was guaranteed to be using a value earlier or the same as what GetAndDecayPressure saved, which seems incorrect (albeit unlikely to matter).